requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
datasets==4.5.0
deepspeed==0.18.5
dill==0.4.0
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
flake8==7.3.0
flash_attn_3==3.0.0b1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-generativeai>=0.8.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
keras==3.12.1
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai>=1.0.0
opt_einsum==3.4.0
optree==0.18.0
packaging>=24.0
pandas==2.3.3
pathspec==1.0.4
peft==0.18.1
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.0
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
sentencepiece==0.2.1
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
tqdm==4.67.1
transformers==5.0.0
triton==3.6.0
trl==0.27.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wandb==0.24.1
Werkzeug==3.1.5
wrapt==2.1.0
xxhash==3.6.0
yarl==1.22.0
